divided hessian code into new code architecture
