+++ batchglm/train/tf/base_glm/README.md
@@ -1,0 +1,2 @@
+# Classes with GLM specific code.
+All noise models that are in the GLM category inherit all of these classes.
