+++ batchglm/train/tf/base_glm/__init__.py
@@ -1,2 +1,3 @@
 from .estimator_graph import GradientGraphGLM, NewtonGraphGLM, TrainerGraphGLM, EstimatorGraphGLM, FullDataModelGraphGLM
+from .hessians import HessiansGLM
 from .model import ESTIMATOR_PARAMS, ProcessModelGLM, ModelVarsGLM, BasicModelGraphGLM
