+++ batchglm/train/tf/base_glm/external.py
@@ -1,6 +1,9 @@
 import batchglm.train.tf.ops as op_utils
 import batchglm.train.tf.train as train_utils
 
+
 from batchglm.models.nb_glm import AbstractEstimator
 
 from batchglm.train.tf.base import ProcessModelBase, TFEstimatorGraph
+
+from batchglm import pkg_constants
