+++ batchglm/train/tf/base_glm_all/__init__.py
@@ -1,2 +1,3 @@
 from .estimator import EstimatorAll
 from .estimator_graph import EstimatorGraphAll
+from .hessians import HessianTF
