--- batchglm/unit_test/__init__.py
@@ -1,8 +1,0 @@
-from .glm_all import Test_Accuracy_GLM_NB
-from .glm_all import Test_AccuracyAnalytic_GLM_NB
-from .glm_all import Test_AccuracySizeFactors_GLM_NB
-from .glm_all import Test_Graph_GLM_NB
-from .glm_all import Test_DataTypes_GLM_NB
-from .glm_all import Test_ExtremValues_GLM_NB
-from .glm_all import Test_Hessians_GLM_NB
-from .glm_all import Test_Jacobians_GLM_NB
