--- batchglm/unit_test/glm_all/__init__.py
@@ -1,8 +1,0 @@
-from .test_acc_glm_all import Test_Accuracy_GLM_NB
-from .test_acc_analytic_glm_all import Test_AccuracyAnalytic_GLM_NB
-from .test_acc_sizefactors_glm_all import Test_AccuracySizeFactors_GLM_NB
-from .test_graph_glm_all import Test_Graph_GLM_NB
-from .test_data_types_glm_all import Test_DataTypes_GLM_NB
-from .test_extreme_values_glm_all import Test_ExtremValues_GLM_NB
-from .test_hessians_glm_nb import Test_Hessians_GLM_NB
-from .test_jacobians_glm_nb import Test_Jacobians_GLM_NB
